Moved maxSpeed in Parameters, base destructor for Flight, wrote randomizer in float, changed all double into float

+++ graphics.cpp
@@ -1,7 +1,6 @@
 #include "graphics.hpp"
 
 sf::Vector2<float> bd::boidToShape(bd::Boid& b) {
-  sf::Vector2<float> v = {static_cast<float>(b.get_Pos()[0]),
-                          static_cast<float>(b.get_Pos()[1])};
+  sf::Vector2<float> v = {b.get_Pos()[0], b.get_Pos()[1]};
   return v;
 }
